# Store the hash of contract code