--- a/core/wasm/src/cache.rs
+++ b/core/wasm/src/cache.rs
@@ -2,7 +2,7 @@ use cached::SizedCache;
 use wasmer_runtime;
 
 use crate::prepare;
-use crate::types::{Config, Error};
+use crate::types::{Config, ContractCode, Error};
 use primitives::hash::hash;
 use primitives::serialize::Encode;
 
@@ -14,12 +14,12 @@ cached_key! {
     MODULES: SizedCache<String, Result<wasmer_runtime::Module, Error>> = SizedCache::with_size(CACHE_SIZE);
     Key = {
         format!("{}:{}",
-            hash(code),
+            code.get_hash(),
             hash(&config.encode().expect("encoding of config shouldn't fail")),
         )
     };
 
-    fn compile_cached_module(code: &[u8], config: &Config) -> Result<wasmer_runtime::Module, Error> = {
+    fn compile_cached_module(code: &ContractCode, config: &Config) -> Result<wasmer_runtime::Module, Error> = {
         let prepared_code = prepare::prepare_contract(code, config).map_err(Error::Prepare)?;
 
         wasmer_runtime::compile(&prepared_code)