#886 
before:
```
test runtime_send_money                     ... bench:     170,030 ns/iter (+/- 12,920)
test runtime_wasm_bad_code                  ... bench:   1,040,480 ns/iter (+/- 147,577)
test runtime_wasm_benchmark_10_reads_legacy ... bench:     883,079 ns/iter (+/- 115,454)
test runtime_wasm_benchmark_storage_100     ... bench:   1,778,676 ns/iter (+/- 235,356)
test runtime_wasm_benchmark_storage_1000    ... bench:  12,418,531 ns/iter (+/- 1,074,537)
test runtime_wasm_benchmark_sum_1000        ... bench:     762,753 ns/iter (+/- 341,578)
test runtime_wasm_benchmark_sum_1000000     ... bench:  13,355,861 ns/iter (+/- 861,692)
test runtime_wasm_set_value                 ... bench:     750,164 ns/iter (+/- 100,792)
```
after:
```
test runtime_send_money                     ... bench:     164,393 ns/iter (+/- 9,350)
test runtime_wasm_bad_code                  ... bench:     462,570 ns/iter (+/- 94,355)
test runtime_wasm_benchmark_10_reads_legacy ... bench:     635,209 ns/iter (+/- 54,746)
test runtime_wasm_benchmark_storage_100     ... bench:   1,489,376 ns/iter (+/- 151,567)
test runtime_wasm_benchmark_storage_1000    ... bench:  11,913,450 ns/iter (+/- 1,179,097)
test runtime_wasm_benchmark_sum_1000        ... bench:     512,183 ns/iter (+/- 91,926)
test runtime_wasm_benchmark_sum_1000000     ... bench:  12,842,724 ns/iter (+/- 1,465,584)
test runtime_wasm_set_value                 ... bench:     503,174 ns/iter (+/- 44,961)
```